api/pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Crhistian Caraballo <cristianjcj35@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
structlog = "^22.3.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.151.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
